+++ app/src/main/java/com/marcogn/kartlog/di/DatabaseModule.kt
@@ -6,6 +6,7 @@ import com.marcogn.kartlog.data.local.DATABASE_NAME
 import com.marcogn.kartlog.data.local.KartLogDatabase
 import com.marcogn.kartlog.data.local.MIGRATION_1_2
 import com.marcogn.kartlog.data.local.MIGRATION_2_3
+import com.marcogn.kartlog.data.local.MIGRATION_3_4
 import com.marcogn.kartlog.data.local.dao.BackupDao
 import com.marcogn.kartlog.data.local.dao.ConsigliamiDao
 import com.marcogn.kartlog.data.local.dao.MedallionsDao
@@ -29,7 +30,7 @@ object DatabaseModule {
     @Singleton
     fun provideDatabase(@ApplicationContext context: Context): KartLogDatabase =
         Room.databaseBuilder(context, KartLogDatabase::class.java, DATABASE_NAME)
-            .addMigrations(MIGRATION_1_2, MIGRATION_2_3)
+            .addMigrations(MIGRATION_1_2, MIGRATION_2_3, MIGRATION_3_4)
             .build()
 
     @Provides